fix: use user timezone for conference expiration date (#91)
Up to now, we used server timezone everywhere (set to 'Europe/Paris' by default).
This means that conferences booked for a give day expired at 23:59, Paris time, on the given day. This could cut off a conference earlier than a user expected.

This PR gets the user timezoneOffset (e.g. for GMT+1, timezoneOffset is -60), stores it in the loginToken table, and uses it to set an expiration date in the right timezone.

Known issue : we use timezoneOffset rather than timezone (e.g. Europe/Paris), so when we change from summer to winter time, conferences booked before the time change that take place after the time change will end at 11pm or 1am, rather than midnight. I ignored this. :)

Left for later : display all times in the user's timezone, in the interface and the emails. This is trickier than expected, and with current featureset (conferences booked for whole day) it's only used for logintoken expiration dates. I would rather push this fix to prod fast since it can cut off conferences for users.

Edit : to use date-fns, we would have to use the name of the timezone (e.g. 'Europe/Paris'), and it would handle the change to summer/winter time.
But the JS native way of getting the timezone is not supported on IE : https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat/resolvedOptions
So I am not using it.