Add a config.HOSTNAME var to avoid host header attacks (#117)
Add a HOSTNAME_WITH_PROTOCOL config var, that will be used to send and display links, instead of using the host header from the request.
This avoids host header poisoning attacks (detected by CodeQL : https://github.com/betagouv/audioconf/security/code-scanning/11?query=ref%3Arefs%2Fheads%2Fmain)

By default, HOSTNAME_WITH_PROTOCOL is http://localhost:${PORT}

We use HOSTNAME_WITH_PROTOCOL (instead of HOSTNAME and PROTOCOL) separately, because that's the var that scalingo review apps have. So we use it for all scalingo apps for simplicity.

This PR adds a scalingo.json file (with which we configure the HOSTNAME_WITH_PROTOCOL vars). When this file is present, the app.json file is ignored, so this PR removes it.

Co-authored-by: Julien DAUPHANT <jdauphant@users.noreply.github.com>